EVOLUZIONE TEMP

+++ code/f_function.py
@@ -402,10 +402,10 @@ def Corr_Evolution(Proj_Psi0,E,V,t,Base_NumRes,Base_Corr):
 def time_tab(t_i,t_f,Nstep,Lo_li):
 	
 	if Lo_li == 0:
-		t_tab = np.insert(10**-5,1,np.linspace(t_i,t_f,num=Nstep))
+		t_tab = np.insert(10**-5,1,np.linspace(t_i,t_f, num=Nstep, dtype=float))
 	else:
-		t_tab = np.insert(10**-5,1,np.logspace(np.log10(t_i),np.log10(t_f),num=Nstep))
+		t_tab = np.insert(10**-5,1,np.logspace(np.log10(t_i),np.log10(t_f), num=Nstep, dtype=float))
 
-	return float(t_tab)
+	return t_tab
